Modularized display()
