+++ testBox2DSDL/display.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "objects.h"
+#include "InitialSetup.h"
+
+void display();
