+++ testBox2DSDL/objects.cpp
@@ -1,0 +1,5 @@
+#include "objects.h"
+
+Square square;
+Triangle triangle;
+Circle circle;
