+++ testBox2DSDL/objects.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "Square.h"
+#include "Triangle.h"
+#include "Circle.h"
+
+extern Square square;
+extern Triangle triangle;
+extern Circle circle;
