+++ testBox2DSDL/testBox2DSDL.vcxproj
@@ -143,7 +143,9 @@
     <ClCompile Include="Car.cpp" />
     <ClCompile Include="Circle.cpp" />
     <ClCompile Include="Constants.cpp" />
+    <ClCompile Include="display.cpp" />
     <ClCompile Include="InitialSetup.cpp" />
+    <ClCompile Include="objects.cpp" />
     <ClCompile Include="Source.cpp" />
     <ClCompile Include="Square.cpp" />
     <ClCompile Include="Triangle.cpp" />
@@ -153,7 +155,9 @@
     <ClInclude Include="Car.h" />
     <ClInclude Include="Circle.h" />
     <ClInclude Include="Constants.h" />
+    <ClInclude Include="display.h" />
     <ClInclude Include="InitialSetup.h" />
+    <ClInclude Include="objects.h" />
     <ClInclude Include="Shape.h" />
     <ClInclude Include="Square.h" />
     <ClInclude Include="Triangle.h" />
